Add with redux HOC to more neatly connect any component
