+++ src/containers/RunningPage.js
@@ -1,12 +1,9 @@
 import React, { Component } from 'react';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 import Countdown from '../components/Countdown';
 import shortBeep from '../assets/shortAlert.mp3';
 import longBeep from '../assets/longAlert.mp3';
-
-import * as actions from '../actions/actions';
+import { withRedux } from '../utils/WithRedux';
 
 class RunningPage extends Component {
     constructor(props) {
@@ -113,14 +110,4 @@ class RunningPage extends Component {
     }
 }
 
-const mapStateToProps = state => ({
-    exercises: state.config.exercises,
-    workTime: state.config.workTime,
-    restTime: state.config.restTime,
-    sets: state.config.sets,
-    isConfigValid: state.isConfigValid
-});
-
-const mapDispatchToProps = dispatch => bindActionCreators(actions, dispatch);
-
-export default connect(mapStateToProps, mapDispatchToProps)(RunningPage);
+export default withRedux(RunningPage);
